HO: MNCC interface uses lchan_mode and lchan_type to modify channel
The MNCC application must provide the lchan_type with the lchan_mode,
since FR and HR codec depends on given lchan_type. In case of AMR, the
lchan_type is ignored and an available channel type is selected. In case
of EFR, the given lchan_type (must be full rate) is used.