BS11: send MRPCI message after HANDOVER COMPLETE and ASSIGNMENT COMPLETE
This message is required for each activated channel. It is sent after
activating channel, due to IMMEDIATE ASSIGNMENT already (after paging
response or CM service request). Also it is required for the activated
channel after late assignment or handover.

It has been successfully tested with late assignment and handover to the
same BTS.