Reduce T3113 default from 60s to 10s
Our T3113 timer default of 60s was set early in the development of
OpenBSC, where we didn't really know what values to use and used
excessively large/safe values. Paging the same MS for 60 seconds (even
if there's no paging response) will however create a lot of PCH load for
no good reason.

It seems there's no clear guidance as to what the value should be. Other
implementations use something in the order of 10 seconds (OpenBTS,
yateBTS), which seems more realistic. THe Siemens BS-11 has a default of
5 seconds.

Let's be conservative and go to 10s as a default, which is already 6
times less potential PCH usage than our default so far...

Closes: OS#2756
Change-Id: If9c8441939c6fdcf6e2b9ede8cc576eb86296209