handover_logic.c: always do inter-cell channel activation
In commit [1], the channel activation kind was modified out-of-context to use
Immediate Assignment in case of intra-cell re-assignment. This so far hasn't
worked out in tests, so instead stay with the strategy to use
RSL_ACT_INTER_ASYNC also for re-assignment within the same cell. We may move to
RSL_ACT_INTRA_IMM_ASS in a separate patch after proper testing.

[1]: 5eaa4fb821751b70ad2b864d8765ff96bd301996,
     I2d91765c1f9719c64fd99426a872cecc724215db
     "HO: add indicators for inter-cell and async ho, use for chan act type",

Note that so far, no re-assignment within cells is triggered by the current
handover decision algorithm, so the effect will be really noticeable with
handover decision 2 that is coming up.

See feature OS#2949 which asks for implementing actual RR Assignment to
handover within a cell.

Related: OS#2949
Change-Id: Ia3d979094fa16620f1acc5933418797829710ba8