flush paging when RSL link is dropped
Flush the paging queue if the link to TRX 0 is dropped from a BTS.
This should prevent stale paging requests sent to the BTS when it
disconnects or reconnects, as seen in the TTCN3 BSC_Tests test suite.

Also, add entries to the log when RSL or OML links are dropped so
that related error messages in the log can be interpreted in context.

Change-Id: If4401c1139cd01faf5ff374301a9a701898c3777
Related: OS#2901