properly skip paging is OML link is down
If the OML link is found down while a paging request is issued,
no paging message is sent. However, we were still counting such
pagings as an actual attempt, and counted them towards the number
of available slots on the paging request queue.

Move the OML link check to the caller of page_ms() where the
accounting steps can be properly skipped as they should be.

Change-Id: I5b6db681da7d45c49e1f2f99d7789c8a29372ef3
Related: OS#2901