fix handling of invalid pchan names in vty
If an invalid phys_chan_config is specified in osmo-bsc.cfg, raise
a parsing error at program startup.

If an invalid phys_chan_config is specified in in the VTY while
the program is running, show a warning to inform the user that
the configuration change was not applied.

Change-Id: I97baa359464a0e94de2497bc9214b99ed2a24041
Related: OS#1876