chan_alloc: reset rtp voice related bits in lchan_free()
The function lchan_free() is supposed to reset the lchan so that
it can be used by another connection. This function does not yet
reset the struct memebers in lchan->abis_ip. This may lead to
confusion if some other end re-uses that lchan and finds old RTP
voice port/ip settings there. Those data must be reset to
ensure it does accidently migrate into an unrelated conext.

- do a memset to 0 on lchan->abis_ip in lchan_free()

Change-Id: I0c99494292cd1d058a19a21413d0ddb51471c6be
Related: OS#3396