assignment: remove unnecessary call to gsm0808_speech_codec_from_chan_type()
There is a function call to gsm0808_speech_codec_from_chan_type() after
an if construct that checks if we need voice and if we need it if it is
related to an AoIP connection.

However, at the moment we call gsm0808_speech_codec_from_chan_type() all
the time, even when just figured out that we don't need to. Presumably
this is a refactory leftover. Lets remove the excess function call.

Change-Id: I34d9281944b36cd89ad8e1c5774f0ea80fdfadc8