inter-BSC HO incoming: send BSSMAP HO Complete directly
Before Handover is fully completed, the gscon remains in an INIT state. To send
back the BSSMAP Handover Complete message, use osmo_bsc_sigtran_send() directly
to not thwart the message due to the gscon state.

(The gscon state will change to ACTIVE right after that, once the handover FSM
is done.)

Change-Id: Ic48ae2bb23565015d5e2ccb56308fad09347b51a