CBCH: Fix gsm_bts_get_cbch()
When the timeslot FSMs were introduced, the function gsm_bts_get_cbch()
was ported to use 'pchan_is'.  However, the function is called very
early, before pchan_is is actually initialized.

Let's make sure we're using the _configured_ channel combination
when resolving where the CBCH is (if any) in the BTS.

Ever since merging the timeslot FSMs and before this patch, SI4 would
never indicate the presence of a CBCH.

Change-Id: I6251b5f3e1180ad466e9349a1845e1b91f661c0b
Related: OS#3532