neighbor vty: identify neighbors by gsm0808 cell id
For deleting neighbor entries and for triggering manual handovers, it is useful
to identify neighbors by the gsm0808_cell_id, instead of just by ARFCN+BSIC.

This is not a hard requirement, since BTS+ARFCN+BSIC fully identifies a
neighbor. But consider these contrived examples of the situation before this
patch:

 # bts 0
 # neighbor cgi 001 01 23 42
 # no neighbor cgi 001 01 23 42
   % No such command
 # no neighbor bts 5
   (manually looked up that BTS 5 has above CGI)

 # neighbor lac 42 arfcn 23 bsic 5
 # do handover any to lac 42
   % No such command
 # do handover any to arfcn 23 bsic 5
   (if multiple cells have the same ARFCN+BSIC, the identification is not unique)

So this fills a gap, if only to help debugging / analyzing handover operations.

Change-Id: I198fe7055d1e09b128693eb51276e3d8cde1c0ba